@@ -1,24 +1,21 @@
 {
   "name": "node-controllab",
-  "version": "1.1.1",
+  "version": "2.0.0",
   "description": "A Javascript module to interface with the LEGO DACTA Control Lab.",
   "main": "dist/index.js",
   "scripts": {
-    "build": "tslint -c tslint.json \"./src/**/*.ts\" && tsc",
+    "build": "tsc",
     "all": "npm run build",
     "prepublishOnly": "npm run build"
   },
   "author": "Nathan Kellenicki <nathan@kellenicki.com>",
   "license": "MIT",
   "dependencies": {
-    "debug": "^4.1.1",
-    "serialport": "^9.0.0"
+    "debug": "^4.4.3",
+    "serialport": "^13.0.0"
   },
   "devDependencies": {
-    "@types/debug": "^4.1.5",
-    "@types/node": "^14.0.5",
-    "@types/serialport": "^8.0.1",
-    "tslint": "^6.1.2",
-    "typescript": "^3.9.3"
+    "@types/node": "^25.0.2",
+    "typescript": "^5.9.3"
   }
 }
