@@ -1,262 +1,303 @@
-import { EventEmitter} from "events";
+import { EventEmitter } from "events";
+import { ControlLabConnection, ControlLabConnectionOptions } from "./connection";
+import { ControlLabState, OUTPUT_PORT_INDICES, OutputPortId, SensorType, TouchEvent } from "./constants";
+import { SensorNotification, createOutputPowerCommand } from "./protocol";
 
-import SerialPort from "serialport";
+const INPUT_PORT_COUNT = OUTPUT_PORT_INDICES.length;
 
-import * as Consts from "./consts";
+interface BaseSensorPayload {
+  inputPort: number;
+  rawValue: number;
+}
 
-const HANDSHAKE_OUTBOUND = "p\0###Do you byte, when I knock?$$$";
-const HANDSHAKE_INBOUND = "###Just a bit off the block!$$$";
-const SENSOR_MESSAGE_LENGTH = 19;
-const SENSOR_MESSAGE_OFFSETS = [14, 10, 6, 2, 16, 12, 8, 4];
-const OUTPUT_INDICES = ["A", "B", "C", "D", "E", "F", "G", "H"];
+export interface TouchSensorPayload extends BaseSensorPayload {
+  kind: "touch";
+  event: TouchEvent;
+  pressed: boolean;
+  force: number;
+}
 
+export interface TemperatureSensorPayload extends BaseSensorPayload {
+  kind: "temperature";
+  fahrenheit: number;
+  celsius: number;
+}
+
+export interface LightSensorPayload extends BaseSensorPayload {
+  kind: "light";
+  intensity: number;
+}
+
+export interface RotationSensorPayload extends BaseSensorPayload {
+  kind: "rotation";
+  rotations: number;
+  delta: number;
+}
+
+export type ControlLabSensorPayload =
+  | TouchSensorPayload
+  | TemperatureSensorPayload
+  | LightSensorPayload
+  | RotationSensorPayload;
+
+export interface ControlLabEventMap {
+  connected: () => void;
+  disconnected: () => void;
+  error: (error: Error) => void;
+  notification: (payload: ControlLabSensorPayload) => void;
+  touch: (payload: TouchSensorPayload) => void;
+  temperature: (payload: TemperatureSensorPayload) => void;
+  light: (payload: LightSensorPayload) => void;
+  rotation: (payload: RotationSensorPayload) => void;
+}
 
 export class ControlLab extends EventEmitter {
+  public state: ControlLabState = ControlLabState.NotReady;
 
+  private readonly connection: ControlLabConnection;
+  private readonly sensorValues = new Array<number>(INPUT_PORT_COUNT).fill(0);
+  private readonly rotationValues = new Array<number>(INPUT_PORT_COUNT).fill(0);
+  private readonly sensorTypes = new Array<SensorType>(INPUT_PORT_COUNT).fill(SensorType.Unknown);
+  private readonly handleSensorNotificationBound: (notification: SensorNotification) => void;
+  private readonly lastSensorPayloads = new Map<string, ControlLabSensorPayload>();
 
-    public state: Consts.State = Consts.State.NOT_READY;
+  constructor(path: string, options?: ControlLabConnectionOptions) {
+    super();
+    this.handleSensorNotificationBound = (notification: SensorNotification) => {
+      this.handleSensorNotification(notification);
+    };
+    this.connection = new ControlLabConnection(path, options);
+    this.connection.on("ready", () => {
+      this.state = ControlLabState.Ready;
+      this.emit("connected");
+    });
+    this.connection.on("disconnect", () => {
+      this.state = ControlLabState.NotReady;
+      this.emit("disconnected");
+    });
+    this.connection.on("notification", this.handleSensorNotificationBound);
+    this.connection.on("error", (error) => this.emit("error", error));
+  }
 
-    private _serialPort: SerialPort | undefined;
-    private _serialPath: string;
-    private _messageBuffer: Buffer = Buffer.alloc(0);
-    private _keepAlive: NodeJS.Timer | undefined;
+  public override on<U extends keyof ControlLabEventMap>(event: U, listener: ControlLabEventMap[U]): this;
+  public override on(event: string | symbol, listener: (...args: unknown[]) => void): this;
+  public override on(event: string | symbol, listener: (...args: unknown[]) => void): this {
+    return super.on(event, listener);
+  }
 
-    private _sensorValues: number[] = new Array(8).fill(0);
-    private _rotationValues: number[] = new Array(8).fill(0);
-    private _forceValues: number[] = new Array(8).fill(0);
+  public override once<U extends keyof ControlLabEventMap>(event: U, listener: ControlLabEventMap[U]): this;
+  public override once(event: string | symbol, listener: (...args: unknown[]) => void): this;
+  public override once(event: string | symbol, listener: (...args: unknown[]) => void): this {
+    return super.once(event, listener);
+  }
 
-    private _connectCallback: any;
+  public override emit<U extends keyof ControlLabEventMap>(event: U, ...args: Parameters<ControlLabEventMap[U]>): boolean;
+  public override emit(event: string | symbol, ...args: unknown[]): boolean;
+  public override emit(event: string | symbol, ...args: unknown[]): boolean {
+    return super.emit(event, ...args);
+  }
 
-    private _sensorTypes: Consts.SensorType[] = new Array(8).fill(Consts.SensorType.UNKNOWN);
+  public async connect(): Promise<void> {
+    await this.connection.open();
+  }
 
+  public async disconnect(): Promise<void> {
+    await this.connection.disconnect();
+  }
 
-    constructor (path: string) {
-        super();
-        this._serialPath = path;
+  public setSensorType(inputPort: number, type: SensorType): void {
+    this.ensureInputPort(inputPort);
+    const index = inputPort - 1;
+    this.sensorTypes[index] = type;
+    this.sensorValues[index] = 0;
+    this.resetRotation(inputPort);
+  }
+
+  public setTouchSensor(inputPort: number): void {
+    this.setSensorType(inputPort, SensorType.Touch);
+  }
+
+  public setTemperatureSensor(inputPort: number): void {
+    this.setSensorType(inputPort, SensorType.Temperature);
+  }
+
+  public setLightSensor(inputPort: number): void {
+    this.setSensorType(inputPort, SensorType.Light);
+  }
+
+  public setRotationSensor(inputPort: number): void {
+    this.setSensorType(inputPort, SensorType.Rotation);
+  }
+
+  public resetRotation(inputPort: number): void {
+    this.ensureInputPort(inputPort);
+    this.rotationValues[inputPort - 1] = 0;
+  }
+
+  public setPower(outputPort: OutputPortId | string, power: number): void {
+    if (!this.connection.isReady) {
+      throw new Error("Control Lab not connected");
     }
-
-
-    public connect (callback?: any) {
-        return new Promise((resolve) => {
-            this._serialPort = new SerialPort(this._serialPath, { baudRate: 9600 });
-            this._serialPort.on("data", this._handleIncomingData.bind(this));
-            this._sendHandshake();
-            this._connectCallback = () => {
-                this._keepAlive = this._startKeepAlive();
-                if (callback) {
-                    callback();
-                }
-                return resolve();
-            }
-        });
+    if (power < -8 || power > 8) {
+      throw new Error("Power must be between -8 and 8");
     }
+    const outputMask = this.getOutputPortMask(outputPort);
+    this.connection.send(createOutputPowerCommand(outputMask, power));
+  }
 
+  public sleep(delay: number): Promise<void> {
+    return new Promise((resolve) => {
+      setTimeout(resolve, delay);
+    });
+  }
 
-    public disconnect() {
-        if (this._serialPort) {
-            this._serialPort.close();
-            this._serialPort = undefined;
-            if (this._keepAlive) {
-                clearInterval(this._keepAlive);
-                this._keepAlive = undefined;
-            }
+  private handleSensorNotification(notification: SensorNotification): void {
+    const payloads: ControlLabSensorPayload[] = [];
+    notification.samples.forEach((sample) => {
+      const inputPort = sample.inputPort;
+      if (!inputPort || inputPort > INPUT_PORT_COUNT) {
+        return;
+      }
+      const index = inputPort - 1;
+      this.rotationValues[index] += sample.rotationDelta;
+      const sensorType = this.sensorTypes[index];
+      switch (sensorType) {
+        case SensorType.Touch: {
+          const pressedValue = sample.rawValue < 1000 ? TouchEvent.Pressed : TouchEvent.Released;
+          if (pressedValue !== this.sensorValues[index]) {
+            this.sensorValues[index] = pressedValue;
+            const force = Math.max(0, Math.min(100, Math.floor(100 - (sample.rawValue / 1024) * 100)));
+            payloads.push({
+              kind: "touch",
+              inputPort,
+              rawValue: sample.rawValue,
+              event: pressedValue,
+              pressed: pressedValue === TouchEvent.Pressed,
+              force
+            });
+          }
+          break;
         }
+        case SensorType.Temperature: {
+          const fahrenheit = +(((760 - sample.rawValue) / 4.4 + 32).toFixed(2));
+          const celsius = +(((760 - sample.rawValue) / 4.4) * (5 / 9)).toFixed(2);
+          if (fahrenheit !== this.sensorValues[index]) {
+            this.sensorValues[index] = fahrenheit;
+            payloads.push({
+              kind: "temperature",
+              inputPort,
+              rawValue: sample.rawValue,
+              fahrenheit,
+              celsius
+            });
+          }
+          break;
+        }
+        case SensorType.Light: {
+          const intensity = Math.floor(146 - sample.rawValue / 7);
+          if (intensity !== this.sensorValues[index]) {
+            this.sensorValues[index] = intensity;
+            payloads.push({
+              kind: "light",
+              inputPort,
+              rawValue: sample.rawValue,
+              intensity
+            });
+          }
+          break;
+        }
+        case SensorType.Rotation: {
+          const rotations = this.rotationValues[index];
+          if (rotations !== this.sensorValues[index]) {
+            this.sensorValues[index] = rotations;
+            payloads.push({
+              kind: "rotation",
+              inputPort,
+              rawValue: sample.rawValue,
+              rotations,
+              delta: sample.rotationDelta
+            });
+          }
+          break;
+        }
+        default:
+          break;
+      }
+    });
+
+    payloads.forEach((payload) => {
+      if (!this.shouldEmitSensorPayload(payload)) {
+        return;
+      }
+      this.emit("notification", payload);
+      switch (payload.kind) {
+        case "touch":
+          this.emit("touch", payload);
+          break;
+        case "temperature":
+          this.emit("temperature", payload);
+          break;
+        case "light":
+          this.emit("light", payload);
+          break;
+        case "rotation":
+          this.emit("rotation", payload);
+          break;
+        default:
+          break;
+      }
+    });
+  }
+
+  private getOutputPortMask(outputPort: OutputPortId | string): number {
+    const normalized = outputPort.toUpperCase();
+    const index = OUTPUT_PORT_INDICES.indexOf(normalized as OutputPortId);
+    if (index === -1) {
+      throw new Error(`Unknown output port '${outputPort}'`);
     }
+    return 1 << index;
+  }
 
-
-    public setSensorType (port: number, type: Consts.SensorType) {
-        this._sensorTypes[port - 1] = type;
-        this._sensorValues[port - 1] = 0;
-        this._forceValues[port - 1] = 0;
-        this.resetRotation(port);
+  private ensureInputPort(inputPort: number): void {
+    if (inputPort < 1 || inputPort > INPUT_PORT_COUNT) {
+      throw new Error(`Input port must be between 1 and ${INPUT_PORT_COUNT}`);
     }
+  }
 
-
-    public resetRotation (port: number) {
-        this._rotationValues[port - 1] = 0;
+  private shouldEmitSensorPayload(payload: ControlLabSensorPayload): boolean {
+    const key = this.getPayloadKey(payload);
+    const previous = this.lastSensorPayloads.get(key);
+    if (previous && this.sensorPayloadsEqual(previous, payload)) {
+      return false;
     }
+    this.lastSensorPayloads.set(key, payload);
+    return true;
+  }
 
+  private getPayloadKey(payload: ControlLabSensorPayload): string {
+    return `${payload.kind}:${payload.inputPort}`;
+  }
 
-    public setPower (output: string, power: number) {
-        if (!this._serialPort) {
-            throw new Error("Control Lab not connected");
-        }
-        if (!(power <= 8 && power >= -8)) {
-            throw new Error("Power must be between -8 and 8");
-        }
-        const outputIndex = 1 << OUTPUT_INDICES.indexOf(output);
-        if (power === 0) {
-            const message = Buffer.alloc(2);
-            message[0] = Consts.Commands.POWER_OFF;
-            message[1] = outputIndex;
-            this._serialPort.write(message);
-            return;
-        }
-        const message = Buffer.alloc(6);
-        if (power < 0) {
-            message[0] = Consts.Commands.DIRECTION_RIGHT;
-        } else {
-            message[0] = Consts.Commands.DIRECTION_LEFT;
-        }
-        power = Math.abs(power) - 1;
-        message[1] = outputIndex;
-        message[2] = Consts.Commands.POWER_LEVEL | power;
-        message[3] = outputIndex;
-        message[4] = Consts.Commands.POWER_ON;
-        message[5] = outputIndex;
-        this._serialPort.write(message);
+  private sensorPayloadsEqual(a: ControlLabSensorPayload, b: ControlLabSensorPayload): boolean {
+    if (a.kind !== b.kind || a.inputPort !== b.inputPort) {
+      return false;
     }
-
-
-    public sleep (delay: number) {
-        return new Promise((resolve) => {
-            setTimeout(resolve, delay);
-        });
+    switch (a.kind) {
+      case "touch":
+        return (
+          b.kind === "touch" &&
+          a.event === b.event &&
+          a.pressed === b.pressed &&
+          a.force === b.force
+        );
+      case "temperature":
+        return b.kind === "temperature" && a.fahrenheit === b.fahrenheit && a.celsius === b.celsius;
+      case "light":
+        return b.kind === "light" && a.intensity === b.intensity;
+      case "rotation":
+        return b.kind === "rotation" && a.rotations === b.rotations && a.delta === b.delta;
+      default:
+        return false;
     }
-
-
-    private _sendHandshake () {
-        this._serialPort?.write(Buffer.from(HANDSHAKE_OUTBOUND));
-    }
-
-
-    private _startKeepAlive () {
-        return setInterval(() => {
-            this._serialPort?.write(Buffer.from([0x02]));
-        }, 2000);
-    }
-
-
-    private _handleIncomingData (data?: Buffer) {
-
-        if (data) {
-            if (!this._messageBuffer) {
-                this._messageBuffer = data;
-            } else {
-                this._messageBuffer = Buffer.concat([this._messageBuffer, data]);
-            }
-        }
-
-        if (this._messageBuffer.length <= 0) {
-            return;
-        }
-
-        if (this.state === Consts.State.NOT_READY) {
-            if (this._messageBuffer.length >= HANDSHAKE_INBOUND.length) {
-                const message = this._messageBuffer.slice(0, HANDSHAKE_INBOUND.length);
-                if (message.includes(HANDSHAKE_INBOUND)) {
-                    this._messageBuffer = this._messageBuffer.slice(this._messageBuffer.indexOf(HANDSHAKE_INBOUND) + HANDSHAKE_INBOUND.length);
-                    this.state = Consts.State.READY;
-                    this.emit("connected");
-                    if (this._connectCallback) {
-                        this._connectCallback();
-                    }
-                    if (this._messageBuffer.length > 0) {
-                        this._handleIncomingData();
-                    }
-                }
-            }
-            return;
-        }
-
-        if (this.state === Consts.State.READY) {
-
-            if (
-                this._messageBuffer.length >= SENSOR_MESSAGE_LENGTH
-                && this._messageBuffer[0] === 0x00
-                // && this._messageBuffer[1] === 0x00 // NK Docs say first two bytes of a message should always be 0x00 0x00, but sometimes its 0x00 0x01?!
-            ) {
-                const message = this._messageBuffer.slice(0, SENSOR_MESSAGE_LENGTH);
-                this._messageBuffer = this._messageBuffer.slice(SENSOR_MESSAGE_LENGTH);
-
-                if (!this._verifyMessage(message)) {
-                    this._messageBuffer = Buffer.alloc(0);
-                    return;
-                }
-
-                this._parseMessage(message);
-                if (this._messageBuffer.length > 0) {
-                    this._handleIncomingData();
-                }
-            }
-        }
-
-    }
-
-
-    private _verifyMessage (message: Buffer) {
-        let checksum = 0;
-        for (let byte = 0; byte < message.length; byte++) {
-            checksum += message[byte];
-        }
-        if ((checksum & 0xff) === 0xff) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-
-    private _parseMessage (message: Buffer) {
-        for (let sensor = 0; sensor < 8; sensor++) {
-            const port = sensor + 1;
-
-            const word = message.slice(SENSOR_MESSAGE_OFFSETS[sensor], SENSOR_MESSAGE_OFFSETS[sensor] + 2);
-
-            const value = (word[0] << 2) | ((word[1] >> 6) & 0x03);
-            const state = word[1] & 0x3f;
-            let change = state & 3;
-            if ((state & 4) === 0) {
-                change *= -1;
-            }
-            this._rotationValues[sensor] += change;
-
-            switch (this._sensorTypes[sensor]) {
-                case Consts.SensorType.TOUCH:
-                {
-                    const pressedValue = (value < 1000) ? 1 : 0;
-                    const force = Math.floor(100 - ((value / 1024) * 100));
-                    if (pressedValue !== this._sensorValues[sensor]) {
-                        this._sensorValues[sensor] = pressedValue;
-                        this.emit("touch", port, { event: pressedValue });
-                    }
-                    if (force !== this._forceValues[sensor]) {
-                        this._forceValues[sensor] = force;
-                        this.emit("force", port, { force });
-                    }
-                    break;
-                }
-                case Consts.SensorType.TEMPERATURE:
-                {
-                    const fahrenheit = +(((760 - value) / 4.4 + 32).toFixed(2));
-                    const celsius = +(((760 - value) / 4.4 * 5 / 9).toFixed(2));
-                    if (fahrenheit !== this._sensorValues[sensor]) {
-                        this._sensorValues[sensor] = fahrenheit;
-                        this.emit("temperature", port, { fahrenheit, celsius });
-                    }
-                    break;
-                }
-                case Consts.SensorType.LIGHT:
-                {
-                    const intensity = Math.floor(146 - (value / 7));
-                    if (intensity !== this._sensorValues[sensor]) {
-                        this._sensorValues[sensor] = intensity;
-                        this.emit("light", port, { intensity });
-                    }
-                    break;
-                }
-                case Consts.SensorType.ROTATION:
-                {
-                    const rotations = this._rotationValues[sensor];
-                    if (rotations !== this._sensorValues[sensor]) {
-                        this._sensorValues[sensor] = rotations;
-                        this.emit("rotate", port, { rotations });
-                    }
-                    break;
-                }
-            }
-
-        }
-    }
-
+  }
 }
